+++ backend/tests/conftest.py
@@ -28,11 +28,6 @@ def client_fixture():
     app_client = TestClient(app)
 
     test_settings = Settings(
-        tools={
-            "literature": {
-                "url": "fake_literature_url",
-            },
-        },
         llm={
             "openai_token": "fake_token",
         },
@@ -182,6 +177,17 @@ def dont_look_at_os_environ(monkeypatch):
             monkeypatch.delenv(env_var, raising=False)
 
 
+@pytest.fixture(autouse=True)
+def disable_boto_metadata_lookup(monkeypatch):
+    """When running tests, we don't want to make any requests to AWS.
+
+
+    boto3.client("service_name") will try to load metadata from the internet.
+    We disable this behavior to avoid network calls during tests.
+    """
+    monkeypatch.setenv("AWS_EC2_METADATA_DISABLED", "true")
+
+
 @pytest.fixture(name="test_user_info")
 def get_default_user_id_vlab_project():
     return (
@@ -309,11 +315,6 @@ async def populate_db(db_connection, test_user_info):
 @pytest.fixture(name="settings")
 def settings():
     return Settings(
-        tools={
-            "literature": {
-                "url": "fake_literature_url",
-            },
-        },
         llm={
             "openai_token": "fake_token",
         },